+++ src/pages/about/index.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 
 import Logo from '@/components/Logo'
 import MetaHead from '@/components/layout/MetaHead'
 
 import { PAGE_PATH } from '@/constants/route'
-import { DEVICE_SIZE } from '@/constants/layout'
 import { BLOG } from '@/domain/owner/constant'
 import { NextPageWithLayout } from '@/shared/types/layout'
+import FONT from '@/styles/constants/font'
 
 const About: NextPageWithLayout = () => {
   return (
@@ -18,13 +18,13 @@ const About: NextPageWithLayout = () => {
       />
       <Box>
         <LogoWrapper>
-          <CustomLogo isFold={false} isHighlightInitial />
+          <Logo isFold={false} isHighlightInitial variety="header_1" />
         </LogoWrapper>
         <DescriptionWrapper>
-          <Paragraph>
+          <Description>
             <Keyword>기본기</Keyword>에 가치를 둡니다.
-          </Paragraph>
-          <Paragraph>
+          </Description>
+          <Description>
             <Strong>단단하고 깊은 뿌리</Strong>를 가진 나무는,
             <br />
             폭풍이나 강풍 중에 넘어질 가능성이 적어 <Strong>안정적</Strong>
@@ -35,18 +35,18 @@ const About: NextPageWithLayout = () => {
             <br />
             <Strong>건강하게 오랫동안</Strong> 생태계에 <Strong>도움</Strong>이
             되는 나무로 살아갑니다.
-          </Paragraph>
-          <Paragraph>
+          </Description>
+          <Description>
             <Strong>탄탄한 기본기</Strong>는,
             <br />
             <Strong>넓은 관점</Strong>을 가지게 하고
             <br /> <Strong>깊은 이해</Strong>를 얻을 수 있으며
             <br /> 견고한 기반으로 <Strong>더 빠르게 성장</Strong> 하도록
             선순환을 이룰 수 있기에
-          </Paragraph>
-          <Paragraph>
+          </Description>
+          <Description>
             <Strong>기본기</Strong>에 초점을 두며 시야를 넓혀가고 있습니다.
-          </Paragraph>
+          </Description>
         </DescriptionWrapper>
       </Box>
     </>
@@ -70,35 +70,22 @@ const LogoWrapper = styled.div`
   margin-bottom: 30px;
 `
 
-const MIN_SIZE = `@media screen and (max-width: ${DEVICE_SIZE.mobile}px)`
-
-const CustomLogo = styled(Logo)`
-  ${({ theme }) => css`
-    ${theme.font.header_1};
-
-    ${MIN_SIZE} {
-      ${theme.font.header_2};
-    }
-  `};
-`
-
 const DescriptionWrapper = styled.div`
   display: flex;
   flex-direction: column;
 `
 
-const Paragraph = styled.p`
+const Description = styled.p`
   word-break: keep-all;
-  ${({ theme }) => theme.font.body_1};
+  ${FONT.body_1};
   color: ${({ theme }) => theme.color.grey800};
   margin: 10px 0;
 `
 
 const Strong = styled.strong`
   color: ${({ theme }) => theme.color.primary400};
-  font-weight: 500;
 `
 
 const Keyword = styled(Strong)`
-  font-size: 22px;
+  ${FONT.header_4};
 `
